+++ strbuf.h
@@ -14,5 +14,5 @@ int buffer_init(buffer_t *buf, size_t reserve);
 void buffer_clear(buffer_t *buf);
 static inline void buffer_free(buffer_t *buf) { free(buf->data); }
 
-ssize_t buffer_putc(buffer_t *buf, const char c);
+int buffer_putc(buffer_t *buf, const char c);
 ssize_t buffer_printf(buffer_t *buf, const char *fmt, ...) __attribute__((format (printf, 2, 3)));
